Put all classes under catobot package
