+++ src/main/java/catobot/command/DeleteCommand.java
@@ -1,17 +1,16 @@
-package command;
+package catobot.command;
 
 import catobot.Storage;
 import catobot.Ui;
-import exception.BotException;
-import exception.InvalidCommandException;
-import item.Deadline;
-import item.TaskList;
+import catobot.exception.BotException;
+import catobot.exception.InvalidCommandException;
+import catobot.item.TaskList;
 
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
 
     private String content;
 
-    protected DoneCommand(String content) {
+    protected DeleteCommand(String content) {
         this.content = content;
     }
 
@@ -21,8 +20,8 @@ public class DoneCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws BotException {
         try {
-            int index = Integer.parseInt(content.substring("done".length()).trim());
-            Ui.respond(tasks.completeTask(index));
+            int index = Integer.parseInt(content.substring("delete".length()).trim());
+            Ui.respond(tasks.deleteTask(index));
         } catch(NumberFormatException e) {
             throw new InvalidCommandException();
         }
