+++ src/main/java/catobot/command/DoneCommand.java
@@ -1,16 +1,16 @@
-package command;
+package catobot.command;
 
 import catobot.Storage;
 import catobot.Ui;
-import exception.BotException;
-import exception.InvalidCommandException;
-import item.TaskList;
+import catobot.exception.BotException;
+import catobot.exception.InvalidCommandException;
+import catobot.item.TaskList;
 
-public class DeleteCommand extends Command {
+public class DoneCommand extends Command {
 
     private String content;
 
-    protected DeleteCommand(String content) {
+    protected DoneCommand(String content) {
         this.content = content;
     }
 
@@ -20,8 +20,8 @@ public class DeleteCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws BotException {
         try {
-            int index = Integer.parseInt(content.substring("delete".length()).trim());
-            Ui.respond(tasks.deleteTask(index));
+            int index = Integer.parseInt(content.substring("done".length()).trim());
+            Ui.respond(tasks.completeTask(index));
         } catch(NumberFormatException e) {
             throw new InvalidCommandException();
         }
