+++ src/main/java/catobot/command/Executable.java
@@ -1,9 +1,9 @@
-package command;
+package catobot.command;
 
 import catobot.Storage;
 import catobot.Ui;
-import exception.BotException;
-import item.TaskList;
+import catobot.exception.BotException;
+import catobot.item.TaskList;
 
 public interface Executable {
     void execute(TaskList tasks, Ui ui, Storage storage) throws BotException;
