+++ src/main/java/catobot/exception/EmptyCommandException.java
@@ -1,13 +1,13 @@
-package exception;
+package catobot.exception;
 
 /**
- * EmptyCommandException is thrown when the expected command details are empty.
+ * EmptyCommandException is thrown when the expected catobot.command details are empty.
  */
 public class EmptyCommandException extends BotException {
     /**
      * Creates an EmptyCommandException.
      *
-     * @param item The item where the description is empty.
+     * @param item The catobot.item where the description is empty.
      */
     public EmptyCommandException(String item) {
         super(String.format("OOPS!!! The description of a / an %s cannot be empty.", item));
